+++ backend/modules/extensions/actions/edit_theme_template.php
@@ -73,7 +73,7 @@ class BackendExtensionsEditThemeTemplate extends BackendBaseActionEdit
 
 		// determine if deleting is allowed
 		$deleteAllowed = true;
-		if($this->record['id'] == BackendModel::getModuleSetting($this->getModule(), 'default_template')) $deleteAllowed = false;
+		if($this->record['id'] == BackendModel::getModuleSetting('pages', 'default_template')) $deleteAllowed = false;
 		elseif(count(BackendExtensionsModel::getTemplates()) == 1) $deleteAllowed = false;
 		elseif($inUse) $deleteAllowed = false;
 		elseif(!BackendAuthentication::isAllowedAction('delete_theme_template')) $deleteAllowed = false;
@@ -92,7 +92,7 @@ class BackendExtensionsEditThemeTemplate extends BackendBaseActionEdit
 		$this->frm = new BackendForm('edit');
 
 		// init var
-		$defaultId = BackendModel::getModuleSetting($this->getModule(), 'default_template');
+		$defaultId = BackendModel::getModuleSetting('pages', 'default_template');
 
 		// build available themes
 		$themes = array();
@@ -331,7 +331,7 @@ class BackendExtensionsEditThemeTemplate extends BackendBaseActionEdit
 				$item['data'] = serialize($item['data']);
 
 				// if this is the default template make the template active
-				if(BackendModel::getModuleSetting($this->getModule(), 'default_template') == $this->record['id']) $item['active'] = 'Y';
+				if(BackendModel::getModuleSetting('pages', 'default_template') == $this->record['id']) $item['active'] = 'Y';
 
 				// if the template is in use we can't de-activate it
 				if(BackendExtensionsModel::isTemplateInUse($item['id'])) $item['active'] = 'Y';
@@ -343,7 +343,7 @@ class BackendExtensionsEditThemeTemplate extends BackendBaseActionEdit
 				BackendModel::triggerEvent($this->getModule(), 'after_edit_template', array('item' => $item));
 
 				// set default template
-				if($this->frm->getField('default')->getChecked() && $item['theme'] == BackendModel::getModuleSetting('core', 'theme', 'core')) BackendModel::setModuleSetting($this->getModule(), 'default_template', $item['id']);
+				if($this->frm->getField('default')->getChecked() && $item['theme'] == BackendModel::getModuleSetting('core', 'theme', 'core')) BackendModel::setModuleSetting('pages', 'default_template', $item['id']);
 
 				// update all existing pages using this template to add the newly inserted block(s)
 				if(BackendExtensionsModel::isTemplateInUse($item['id'])) BackendPagesModel::updatePagesTemplates($item['id'], $item['id'], $this->frm->getField('overwrite')->getChecked());