+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Bugfixes:
 * Location: fixed a javascript error with jquery.
 * CSS: fixed Safari bug for DataGrid in Tabs #212
 * Pages: subpages in the footer are now visible in the backend.
+* Pages: default template wasn't used when adding a page.
 
 3.2.3 (2012-01-31)
 --